# openapi-resolved.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Booking System",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8085"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "default"
        ],
        "summary": "Generate Auth Token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "userName": "ADMIN",
                  "password": "Admin@123"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/resources": {
      "get": {
        "tags": [
          "default"
        ],
        "summary": "List Resources (Paginated)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "example": "0"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "example": "10"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "default"
        ],
        "summary": "Create Resource (ADMIN only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "Meeting Room",
                  "type": "Room B1",
                  "description": "Big meeting room",
                  "capacity": 25,
                  "active": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/resources/3": {
      "get": {
        "tags": [
          "default"
        ],
        "summary": "Get Resource by ID",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/resources/4": {
      "put": {
        "tags": [
          "default"
        ],
        "summary": "Update Resource (ADMIN only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "Conference Room B10",
                  "type": "Room",
                  "description": "Updated description",
                  "capacity": 30,
                  "active": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "default"
        ],
        "summary": "Delete Resource (ADMIN only)",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/reservations": {
      "get": {
        "tags": [
          "default"
        ],
        "summary": "List Reservations (with filters + pagination)",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "PENDING"
          },
          {
            "name": "minPrice",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "example": "100"
          },
          {
            "name": "maxPrice",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "example": "5000"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "example": "0"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "example": "5"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "createdOn,desc"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "default"
        ],
        "summary": "Create Reservation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "resourceId": 2,
                  "price": 4500,
                  "startTime": "2025-09-20T14:00:00Z",
                  "endTime": "2025-09-20T16:00:00Z"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/reservations/4": {
      "get": {
        "tags": [
          "default"
        ],
        "summary": "Get Reservation by ID",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "default"
        ],
        "summary": "Update Reservation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "resourceId": 1,
                  "price": 290,
                  "startTime": "2025-09-20T15:00:00Z",
                  "endTime": "2025-09-20T17:00:00Z"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/reservations/5": {
      "delete": {
        "tags": [
          "default"
        ],
        "summary": "Delete Reservation",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {},
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}